# Conflicts:
#	src/edu/utexas/wrap/Link.java

+++ src/edu/utexas/wrap/Priced.java
@@ -1,9 +1,7 @@
 package edu.utexas.wrap;
 
-import java.math.BigDecimal;
-
 public interface Priced {
 
-	public BigDecimal getPrice(Float vot, VehicleClass c);
+	public Double getPrice(Float vot, VehicleClass c);
 
 }
